#258 코드 분할하기
