+++ src/components/18/components/Auth.js
@@ -1,6 +1,6 @@
 import classes from "./Auth.module.css";
 import { useDispatch } from "react-redux";
-import { authActions } from "../store/index";
+import { authActions } from "../store/auth";
 
 const Auth = () => {
   const dispatch = useDispatch();
@@ -11,7 +11,7 @@ const Auth = () => {
     dispatch(authActions.login());
   };
 
-  
+
   return (
     <main className={classes.auth}>
       <section>
